Adding filters to the session

+++ app/Http/Controllers/CompanyController.php
@@ -12,56 +12,88 @@ use Illuminate\Support\Facades\Redirect;
 class CompanyController extends Controller
 {
     public function create()
-    {
-        return view('company/create');
+    {   
+        if (Auth::check()) {
+
+            return view('company/create');
+        } else {
+
+            $companies = Company::all();
+            return View::make("welcome", array("companies" => $companies));            
+        }
     }
+
 
     public function store(Request $request)
     {
-    	$data = [
-            'company' => Input::get('dependencia'),
-            'responsable1' => Input::get('responsable'),
-            'tel1' => Input::get('telefono'),
-            'email1' => Input::get('emailu'),
-            'address' => Input::get('direccionu')
-        ];
-        $input = $request->all();
-        Company::create($data);
         $companies = Company::all();
-        return View::make("admin/hello", array("companies" => $companies));
+
+        if (Auth::check()) {
+        	$data = [
+                'company' => Input::get('dependencia'),
+                'responsable1' => Input::get('responsable'),
+                'tel1' => Input::get('telefono'),
+                'email1' => Input::get('emailu'),
+                'address' => Input::get('direccionu')
+            ];
+            $input = $request->all();
+            Company::create($data);
+            return View::make("admin/hello", array("companies" => $companies));
+        } else {
+
+            return View::make("welcome", array("companies" => $companies));               
+        }
     }
 
     public function edit($id)
     {
-    	$company = Company::findOrFail($id);
-    	return view('company/edit')->withCompany($company);
+        if (Auth::check()) {
+        	$company = Company::findOrFail($id);
+        	return view('company/edit')->withCompany($company);
+        } else {
+
+            $companies = Company::all();
+            return View::make("welcome", array("companies" => $companies));                        
+        }
     }
 
     public function update(Request $request, $id)
     {
-        $company = Company::findOrFail($id);
-	    $data = [
-            'company' => Input::get('dependencia'),
-            'responsable1' => Input::get('responsable'),
-            'tel1' => Input::get('telefono'),
-            'email1' => Input::get('email'),
-            'address' => Input::get('direccion')
-        ];
+        $companies = Company::all();
 
-	    $input = $request->all();
+        if (Auth::check()) {
+            $company = Company::findOrFail($id);
+    	    $data = [
+                'company' => Input::get('dependencia'),
+                'responsable1' => Input::get('responsable'),
+                'tel1' => Input::get('telefono'),
+                'email1' => Input::get('email'),
+                'address' => Input::get('direccion')
+            ];
 
-	    $company->fill($data)->save();
-	    $companies = Company::all();
-        return View::make("admin/hello", array("companies" => $companies));
+    	    $input = $request->all();
+
+    	    $company->fill($data)->save();    	    
+            return View::make("admin/hello", array("companies" => $companies));
+        } else {
+
+            return View::make("welcome", array("companies" => $companies));                           
+        }
     }
 
     public function eliminar($id)
     {
-        $company = Company::findOrFail($id);
-
-        $company->delete();
-
         $companies = Company::all();
-        return View::make("admin/hello", array("companies" => $companies));
+
+        if (Auth::check()) {
+
+            $company = Company::findOrFail($id);
+            $company->delete();
+
+            return View::make("admin/hello", array("companies" => $companies));
+        } else {
+
+            return View::make("welcome", array("companies" => $companies));                           
+        }
     } 
 }
